Separeted recipes logic from lilo stuff
